requirements.txt
requests
pyautoit
opencv-python
numpy
imutils